Re-use the version number information from Cabal.
The BNFC executable claimed to be version "2.6a" -- although the Cabal file
said that it would be version 2.6.0.3. This kind of mismatch can be avoided by
re-using the version number from the Paths_BNFC module which Cabal generates
during the build.

Furthermore, the version number "2.6a" is *illegal* in Cabal. Cabal cannot
parse that string, which means that version 2.6.0.3 of the BNF Converter cannot
be used in Cabal-based builds. An attempt to refer to BNFC as a build tool
invariably ends with this the following message:

 | Warning: cannot determine version of /home/simons/.nix-profile/bin/bnfc :
 | "2.6a\n"
 | setup: The program BNFC version >=2.4 is required but the version of
 | /home/simons/.nix-profile/bin/bnfc could not be determined.

This patch avoids this kind of error.