Change the makefiles slightly for c,c++
This makes it so were not linking against *.o
IT defines a var that has the generic objects in it so that
we can just link those in and link them etc...
I also did some small stuff to make c/c++ more similar.
There is more work needed there that will come in the next commit.