Added OBJS define to Makefiles that are produced
This cleans things up a little on the Makefile and allows us to
make multiple binaries without linking *.o
(If your making a test program and a production program you don't
want to link the test.o to your final program.)

I also worked a little bit on making the files more similar so
it's easier to maintain them.  There is still more work to do here,
but this is a good start.