Exposes writeFiles, makeHaskell and parseCF in the library
The current way of using BNFC is as an executable. However, given that
BNFC is written in Haskell, why not expose its functionalities as
library calls to other Haskell programs?

This is the case of the cal-src project relating to the syntax of the
CAL actor language, which is a cabal-ised project that is a library
that exposes:

    library
      exposed-modules: AbsCAL
                     , LexCAL
         	     , ParCAL
                     , PrintCAL
                     , SkelCAL

The workflow in `Setup.hs` for cal-src is:

1. As a pre-build hook, run `writeFiles` with `makeHaskell` to create
   files AbsCAL.hs, LexCAL.hs, ParCAL.hs, PrintCAL.hs and SkelCAL.hs
   corresponding to the "CAL.cf" grammar file in the cal-src
   repository.

2. Run cabal to create the cal-src library.

The `Setup.hs` for cal-src is:

    main :: IO ()
    main = defaultMainWithHooks simpleUserHooks { preBuild = calGenHook }

    calGenHook :: Args -> BuildFlags -> IO HookedBuildInfo
    calGenHook _ _ = do
      let opts = defaultOptions { lang = "CAL" }
      grammar <- readFile "CAL.cf"
      cf <- parseCF opts TargetHaskell grammar
      writeFiles "." $ makeHaskell opts cf
      return emptyHookedBuildInfo

The move away from calling `bnfc` within Haskell to using the
(proposed) exposed BNFC library is here:

https://github.com/robstewart57/cal-src/commit/c9165159488c19de7393d72c07fd55bd4e4dab31