Add -antlr4 in Java backend. Add JavaLexerParser datatype.
This commit applies the changes and it does not structurally
change the status quo.

Currently the JavaLexerParser type defined in BNFC.Options
is pushed down to other Java specific files of this commit.
This has to be removed for cleansiness, since the
tool-specific features of JLex, JFlex and ANTLR4 can be
decided for once before the code generation begins.