[haskell] Print a space before '.'
The problem is mostly that this was a discrepency between the haskell
pretty printer and the other backends and it was making some test fails.

I changed the Haskell backend instead of the other ones mostly because
that was a smaller change. In addition, I know of '.' being used as
an operator (like in haskell) which would justify spaces before and
after, but I can't think of a language where it is used as a
'terminator'.

At some point we should try to have those things configurable.