Revert "Java Backend: Re-support cup version 0.11a"
This reverts commit a65a2653fb01c3c451cf6e91c25b4ba5ffc2c334.

The location tracking of cup was really awful in 0.11a, so much so that tracking
locations in the abstract syntax tree is basically impossible. Revert this
commit and add a comment about the minimal version required when line numbering
is requested. Verified that cup 0.11a is supported without line numbering.