Java: Add line numbering support to the Absyn files.
Adds line numbers to the generated Absyn files with -l
Adds cup support to record the line numbers as the Absyn tree is built.
When compiled with line number support, the symbolFactory argument must now be
supplied to the parser. All of the examples have been updated to reflect this.
Later cup versions complain that the constructor without a symbol factory is
deprecated. Without line numbering, passing the SymbolFactory to the constructor
doesn't hurt anything.