[ fixed #210 ] generate entrypoints also for CoercCat
Categories with higher precedence can now also be entrypoints.
This was not implemented by BNFC (CF.hs).
While switching it on worked out-of-the-box for the Haskell backend,
it was not supported by the backends:  C, CPP, Java, OCaml.

Musing:
Maybe it had been supported originally, but then was removed.
I found a comment of M.F. from 2004 about fixing the [Ty2] bug,
maybe this disabled CoercCat entrypoints.