[ #163 ] Set Bison max stack size to 10.000.000 (for right recursion)
Default is only 10.000, which seems an anachronism in 2019.

Parsing right-recursive categories needs O(n) stack size, thus,
YYMAXDEPTH is a hard limit on the depth of right recursion.

BNFC attempts to rewrite right recursion to left recursion, but only
when it can be done easily.  Thus, we might still have right recursion
left in the generated Bison grammar.