[ #163 ] Haskell: removed right-to-left recursion "optimization"
For LR parsers that allocate the parse stack in the heap, there is only
a minimal difference between left- and right-recursion.  Right recursion
requires O(n) stack in comparison of O(1) stack of left recursion.
However, the ASTs constructed are the same, and their nodes are stored
in the data stack, which has thus the same size for left- and right
recursion.  Only the control stack is different, but the O(n) extra
machine words are dominated by the size of the ASTs.
Asymptotic space complexity (linear) is certainly the same for both
types of recursion.