Fixed quadratic behavior in C++ generated printer.
This regression was introduced in the fix of issue #242 (commit
a065117a9393d81b1ddb020cba4aa1b40171f80f).  It would cause the C++
generated printer to call resizeBuffer for every string appended to the
buffer, not just when resizing was needed.  So for each call to
bufAppend(char*/String), O(n) bytes were copied, with the complexity
deteriorating to O(n^2).