[ #297 Agda ] AST: preliminary binding for position token
The position parts are bound to a postulated type #Int with postulated
conversions to Nat.

Since position tokens are represented in Haskell rather ugly as

  newtype MyPosId = MyPosId ((Int, Int), String)

we also have to define a type of pairs that we bind to Haskell's tuple
type.

TODO: generate nicer interface to access contents of a position token,
like projections that return line, column, and string.