[ options ] new: --force ignores sanity checks on grammar
Also: only report ERROR if grammar can definitely not be processed by
backend.  It turned out that the previous ERROR reporting was too
strict, e.g. non-unique-names-ignoring-case can be processed by the C++
backends and non-unique-names by the C backend.

Option --force allows to process a grammar with a backend even if the
resulting output is likely ill-formed.