[ close #214 ] File position info with error messages and warnings
Identifiers are now lexed as position tokens, this information is
propagated into various declarations forms (e.g. Rule).

This allows to equip error messages with position information, which
should contribute to the clarity of the error messages, and in
particular make the error for lower case rule names less confusing.

Consider #214 fixed.