[ fixed #314 ] Fix compilation warnings in C backend
Added `-D_POSIX_C_SOURCE=200809L` C compilation flag to make strdup()
and fileno() functions visible in system headers.

This fixes nasty bug where strdup() on x86_64 would return an invalid
pointer because of an invalid implicit declaration i.e. int strdup();
instead of char* strdup();. Invalid value is returned because strdup()
result is treaded as an int (that is 32bit) instead of char* (that is
64bit), thus truncating the upper 32 bits of the pointer.