[ cabal, doctests ] use cabal-doctest; package library BNFC
The current doctests.hs module was a hack since it hard coded
`dist`-paths in order to access the generated LexBNF.hs and ParBNF.hs.

I instead now use the cabal-doctest package that generates a
Build_doctests module with all the necessary information to invoke
doctest.  This saves me from figuring out the paths to the
cabal-generated files.

To get BNFC compiled before doctests (which is necessary to have the
Build_doctests module), I had to package BNFC into a library to depend
on.  As a side benefit, the unit-tests can now depend on this library
rather than recompiling all the modules.