[ #363 ] test case: Norell's example for 'define'
This test has a non-linear definition:
```
define inc x        = Assign x (EOp (EVar x) Plus (EInt 1)) ;
```
This creates syntax trees that not trees bug DAGs.
```
  make_Assign (x, make_EOp (make_EVar(x), make_Plus(), make_EInt(1)))
```
E.g. here, `x` is shared.  This causes problems for the C deallocater
which assumes that it is dealing with a tree.

This issue is left open for now.