Stack CI: make cache keys more fine-grained
We include the commit SHA to ensure that caches always are written.

Note: The stack.yaml does not track the dependencies.
We would need stack.yaml.lock, but don't want this.
We may use the .cabal file instead to get a better approximation.
So far we didn't have any problems, so stick to the current strategy.