Drop support for GHC 8.0 under stack
Because of `cabal-version: 2.0`, stack wants an
`extra-deps: [Cabal-2.0.0.1]` in the `stack-8.0.2.yaml` file, which in
turn requires a `Setup.hs` present even for `build-type: simple`.
(Don't as me, this is stack way of seeing the world.)

I anyway doubt that anyone would want to insists on stack-installing
BNFC with GHC 8.0.
Principle: drop palaeo-GHCs as soon as they cause extra effort.