Fix #453: Java/CUP w/ line numbers: Compatibility with JFlex >= 1.8
Since JFLex 1.8, `yychar` is a `long` rather than an `int` to allow inputs > 2GB.
It seems that JavaCUP has not provided a `ComplexSymbolFactoryLocation(int, int, long)` constructor yet.

Our workaround is to cast `yychar` to `int` when creating a `Location`.
This will produce garbage for inputs > 2GB due to overflow, but this is the situation with JFLex < 1.8 anyway.
A proper fix has to wait for JavaCUP to catch up.