Fix #455: Java/ANTLR: Capital name for Lexer and Parser
Triggered by `example/C.cf`.

ANTLR/Java does not deal well with a parser name `cParser` because with Java 20 `cParserListerner` becomes an unbound symbol.
The lower case `c` is due to the package name sanitization introduced in:
- #212

So this problem exists since 2.8.4, but many only surfaced with recent Java versions that are more strict with case.

A solution is to use a capitial CamelCase for the prefix of `Parser` and `Lexer`, but keep the snake_case for the package name.