Can't add `Lex/Par` to `autogen-modules` in `BNFC.cabal`
These modules should not be shipped but be generated by `alex` and
`happy` **on the target system**.

However, if we add them to `autogen-modules` even the `Lex.x/Par.y` will
not be shipped with `cabal sdist`.

Solution is to make sure that there is no `Lex/Par.hs` along `Lex.x/Par.y`
when invoking `cabal sdist`.  This is embodied in new goal `make sdist`.